\name{example.files}
\alias{example.files}
\title{Downloads the files needed for the example application}
\usage{
example.files()
}
\arguments{
}
\value{
This function returns the vector file for "Caverna do Maroaga (Presidente Figueiredo)" as a sf object. The MapBiomas images are downloaded as a list with each element (of a total of 10) being a raster object.
}
\description{
This function downloads the vector file for "Caverna do Maroaga (Presidente Figueiredo)", which is an Environmentally Protected Area in northern Brazil. The MapBiomas collection 8 cropped for the existing vector file is also downloaded from 2013 to 2022 (10 years worth of data).
}
\examples{
example.files()
}
